requirements.txt
requests
PyYAML
google-cloud-dns
